fix: split aggregate-vectors into loaders module (ROK-949)
Two CI lint errors from the review commit:
- unused 'and' import in aggregate-vectors.ts
- aggregate-vectors.ts was 379 lines (max 300)

Split loaders + groupBy/groupMap helpers into
aggregate-vectors-loaders.ts. Main file now 261 lines; loaders 157.
Also extracted applyInterestSignal helper to keep assembleSignals
under the 30-line function limit.
